+++ src/main/scala/org/broadinstitute/hail/driver/AnnotateVariantsTable.scala
@@ -3,7 +3,6 @@ package org.broadinstitute.hail.driver
 import org.broadinstitute.hail.Utils._
 import org.broadinstitute.hail.annotations.Annotation
 import org.broadinstitute.hail.expr._
-import org.broadinstitute.hail.io.annotators._
 import org.broadinstitute.hail.utils.{ParseContext, ParseSettings}
 import org.broadinstitute.hail.variant.Variant
 import org.kohsuke.args4j.{Argument, Option => Args4jOption}
@@ -25,7 +24,7 @@ object AnnotateVariantsTable extends Command {
       usage = "Specify identifier to be treated as missing")
     var missingIdentifier: String = "NA"
 
-    @Args4jOption(required = false, name = "-v", aliases = Array("--vcolumns"),
+    @Args4jOption(required = false, name = "-k", aliases = Array("--keys"),
       usage = "Specify the column identifiers for chromosome, position, ref, and alt (in that order)")
     var vCols: String = "Chromosome,Position,Ref,Alt"
 
@@ -33,8 +32,9 @@ object AnnotateVariantsTable extends Command {
       usage = "Select only certain columns.  Enter columns to keep as a comma-separated list")
     var select: String = _
 
-    @Args4jOption(required = false, name = "--noheader", usage = "indicate that the file has no header and columns should be indicated by number (0-indexed)")
-    var noheader: Boolean = _
+    @Args4jOption(required = false, name = "--no-header",
+      usage = "indicate that the file has no header and columns should be indicated by `_1, _2, ... _N' (0-indexed)")
+    var noHeader: Boolean = _
 
     @Args4jOption(required = false, name = "-d", aliases = Array("--delimiter"),
       usage = "Field delimiter regex")
@@ -46,14 +46,13 @@ object AnnotateVariantsTable extends Command {
 
     @Argument(usage = "<files...>")
     var arguments: java.util.ArrayList[String] = new java.util.ArrayList[String]()
-
   }
 
   def newOptions = new Options
 
   def name = "annotatevariants table"
 
-  def description = "Annotate variants with TSV file"
+  def description = "Annotate variants with delimited text file"
 
   def supportsMultiallelic = false
 
@@ -88,7 +87,7 @@ object AnnotateVariantsTable extends Command {
       types = Parser.parseAnnotationTypes(options.types) + keySig,
       keyCols = vCols,
       useCols = Option(options.select).map(o => Parser.parseIdentifierList(o)),
-      hasHeader = !options.noheader,
+      noHeader = options.noHeader,
       separator = options.separator,
       missing = options.missingIdentifier,
       commentChar = Option(options.commentChar))
@@ -104,7 +103,8 @@ object AnnotateVariantsTable extends Command {
         val pl = parser(line.value)
         val v = pl.key match {
           case Array(variant) => variant.asInstanceOf[Variant]
-          case Array(chr, pos, ref, alt) => Variant(chr.asInstanceOf[String], pos.asInstanceOf[Int], ref.asInstanceOf[String], alt.asInstanceOf[String])
+          case Array(chr, pos, ref, alt) =>
+            Variant(chr.asInstanceOf[String], pos.asInstanceOf[Int], ref.asInstanceOf[String], alt.asInstanceOf[String])
         }
         (v, pl.value): (Variant, Annotation)
       })